user

+++ src/main/java/com/example/blogapp/sevice/Impl/UserServiceImpl.java
@@ -7,7 +7,11 @@ import com.example.blogapp.entity.User;
 import com.example.blogapp.exception.ApplicationException;
 import com.example.blogapp.mapper.UserMapper;
 import com.example.blogapp.repository.UserRepository;
+import com.example.blogapp.sevice.UserService;
+import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -17,19 +21,22 @@ import java.util.Optional;
 
 @Service
 @AllArgsConstructor
-public class UserSeviceImpl {
+public class UserServiceImpl implements UserService {
     private final UserRepository userRepository;
 
-    public List<UserDto> getUsers() {
-        List<User> users = userRepository.findAll();
-        List<UserDto> userDtos = new ArrayList<>();
-        for (User user : users) {
-            UserDto userDto = UserMapper.map(user);
-            userDtos.add(userDto);
-        }
-        return userDtos;
+    @Override
+    public Page<UserDto> getUsers(Pageable pageable) {
+//        List<User> users = userRepository.findAll();
+//        List<UserDto> userDtos = new ArrayList<>();
+//        for (User user : users) {
+//            userDtos.add(UserMapper.map(user));
+//        }
+//        return userDtos;
+        Page<User> users = userRepository.findAll(pageable);
+        return users.map(UserMapper::map);
     }
 
+    @Override
     public UserDto getUserById(Long id) {
        Optional<User> optionalUser = userRepository.findById(id);
        if (optionalUser.isEmpty()){
@@ -40,10 +47,12 @@ public class UserSeviceImpl {
        return userDto;
     }
 
-    public UserDto createUser(UserCreateForm  userCreateForm) {
-        User user = userRepository.findByEmail(userCreateForm.getEmail());
-        if (user != null){
-            throw new ApplicationException("Email User already exists");
+    @Transactional
+    @Override
+    public UserDto createUser(UserCreateForm userCreateForm) {
+        Optional<User> user = userRepository.findByEmail(userCreateForm.getEmail());
+        if (user.isPresent()) {
+            throw new ApplicationException("Email already exists");
         }
         User users= UserMapper.map(userCreateForm);
         User saveUser= userRepository.save(users);
@@ -51,6 +60,8 @@ public class UserSeviceImpl {
         return  userDto;
     }
 
+    @Transactional
+    @Override
     public UserDto updateUser(UserUpdateForm userUpdateForm, Long id) {
         Optional<User> optionalUser = userRepository.findById(id);
         if (optionalUser.isEmpty()){
@@ -58,10 +69,18 @@ public class UserSeviceImpl {
         }
         User user = optionalUser.get();
         UserMapper.map(userUpdateForm,user);
-        UserDto userDto = UserMapper.map(user);
+        UserDto userDto = UserMapper.map(userRepository.save(user));
         return userDto;
     }
-    public void deleteUser(Long id) {
+
+    @Transactional
+    @Override
+    public String deleteUser(Long id) {
+        Optional<User> optionalUser = userRepository.findById(id);
+        if (optionalUser.isEmpty()){
+            throw new ApplicationException("User not found");
+        }
         userRepository.deleteById(id);
+        return "Delete success";
     }
 }
